CorpusBuilderApp/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
bitsandbytes==0.45.5
black==25.1.0
camelot-py==1.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cryptography
cssselect==1.3.0
cycler==0.12.1
datasets==3.6.0
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
EbookLib==0.19
et_xmlfile==2.0.0
evaluate==0.4.3
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.10.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
kiwisolver==1.4.8
langdetect==1.0.9
lxml==5.4.0
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.41.0
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
numpy==2.1.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parsel==1.10.0
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
peft==0.15.2
pillow==11.0.0
platformdirs==4.3.8
plotly==6.1.1
pluggy==1.5.0
polars==1.29.0
propcache==0.3.1
Protego==0.4.0
protobuf==6.30.2
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
PyDispatcher==2.0.7
PyMuPDF==1.25.5
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.5.0
PySide6>=6.5,<7
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
queuelib==1.8.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
Scrapy==2.13.1
seaborn==0.13.2
selenium==4.33.0
sentence-transformers==4.1.0
sentry-sdk==2.27.0
service-identity==24.2.0
setproctitle==1.3.6
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
# torch==2.8.0.dev20250504+cu128  # (REMOVED: not available for Python 3.12, see below)
torch  # Install the latest compatible version for your Python/CUDA. If you need a specific build, install manually as per your GPU requirements.
torchaudio==2.6.0.dev20250505+cu128
torchvision==0.22.0.dev20250505+cu128
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
w3lib==2.3.1
wandb==0.19.11
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zope.interface==7.2
keyring
pytest-qt
qdarktheme
